* Move theme related functions to Theme.php in the default theme directory. Every theme can control its context easier.
